lambda/requirements.txt
databricks-sql-connector>=2.9.0
